@@ -20,44 +20,45 @@
   "license": "GPL",
   "dependencies": {
     "autoprefixer-loader": "3.1.0",
-    "core-js": "1.1.4",
-    "css-loader": "0.17.0",
-    "diff": "2.1.1",
-    "extract-text-webpack-plugin": "0.8.2",
+    "core-js": "1.2.6",
+    "css-loader": "0.22.0",
+    "diff": "2.2.1",
+    "extract-text-webpack-plugin": "0.9.1",
     "file-loader": "0.8.4",
-    "gulp-util": "3.0.6",
-    "less": "2.5.1",
+    "gulp-util": "3.0.7",
+    "less": "2.5.3",
     "less-loader": "^2.2.1",
     "lodash": "3.10.1",
     "moment": "2.10.6",
-    "node-notifier": "4.2.3",
-    "react": "0.13.3",
-    "react-router": "0.13.3",
+    "node-notifier": "4.3.1",
+    "react": "^0.14.2",
+    "react-dom": "^0.14.2",
+    "react-router": "0.13.5",
     "rimraf": "2.4.3",
-    "style-loader": "0.12.4",
+    "style-loader": "0.13.0",
     "superagent": "1.4.0",
-    "ts-loader": "0.5.5",
+    "ts-loader": "0.7.1",
     "typescript": "1.6.2",
     "url-loader": "0.5.6",
-    "webpack": "1.12.2"
+    "webpack": "1.12.4"
   },
   "devDependencies": {
-    "chai": "3.2.0",
-    "gulp-tslint": "3.2.0",
-    "karma": "0.13.9",
+    "chai": "3.4.1",
+    "gulp-tslint": "3.6.0",
+    "karma": "0.13.15",
     "karma-chai": "0.1.0",
     "karma-chai-sinon": "0.1.5",
-    "karma-chrome-launcher": "0.2.0",
-    "karma-cli": "0.1.0",
+    "karma-chrome-launcher": "0.2.1",
+    "karma-cli": "0.1.1",
     "karma-mocha": "0.2.0",
-    "karma-sourcemap-loader": "0.3.5",
+    "karma-sourcemap-loader": "0.3.6",
     "karma-webpack": "1.7.0",
-    "mocha": "2.2.5",
+    "mocha": "2.3.3",
     "react-hot-loader": "1.3.0",
-    "sinon": "1.16.1",
+    "sinon": "1.17.2",
     "sinon-chai": "2.8.0",
-    "tsd": "0.6.4",
-    "tslint": "2.4.4",
-    "webpack-dev-server": "1.11.0"
+    "tsd": "0.6.5",
+    "tslint": "2.5.1",
+    "webpack-dev-server": "1.12.1"
   }
 }
