@@ -1,3 +1,3 @@
-/// <reference path="tsd.d.ts" />
+/// <reference path="tsd/tsd.d.ts" />
 /// <reference path="versionpress-custom/karma-chai-sinon.d.ts" />
 /// <reference path="versionpress-custom/react-custom.d.ts" />
